consistently use abstract file I/O interfaces
added typesafe H5Part attribute reader to C++ layer
implement H5PartReader with C++ wrappers
added IFileReader and settings file to initializer framework
access I/O only through abstract interface
pass settings map to observables factory
pass abstract reader to initializers
add rank information to file I/O interfaces
pass abstract IFileReader to propagator::load
renamed mpi_file_utils to h5part wrappers
move writeSettings into free function
fix non-HDF5 build
added FileMode to IFileReader interface
read glass templates through IFileReader
added cli util to create parameter files

+++ sph/include/sph/hydro_turb/turbulence_data.hpp
@@ -70,7 +70,7 @@ public:
     //! Normalized solenoidal weight
     T solWeightNorm;
 
-    size_t         numModes;
+    uint64_t       numModes;
     std::vector<T> modes;
     std::vector<T> amplitudes;
     std::vector<T> phases;
